// src/payload-types.ts
export interface Meme {
  id: number;
  name: string;
  slug: string;
  creator?: string | null;
  dateAcquired?: string | null;
  isMemeNft?: boolean | null;
  isToken?: boolean | null;
  isPhysical?: boolean | null;
  knowYourMemeUrl?: string | null;
  provenanceUrl?: string | null;
  description?: string | null;
  image?: (number | null) | Media;
  thumbnail?: (number | null) | Media;
  sortOrder?: number | null;
  updatedAt: string;
  createdAt: string;
  _status?: ('draft' | 'published') | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: number;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "liquid-assets".
 */
export interface LiquidAsset {
  id: number;
  name: string;
  slug: string;
  creator?: string | null;
  ticker?: string | null;
  blockchain?: ('Ethereum' | 'Solana' | 'Base') | null;
  coingeckoUrl?: string | null;
  description?: string | null;
  image?: (number | null) | Media;
  thumbnail?: (number | null) | Media;
  sortOrder?: number | null;
  updatedAt: string;
  createdAt: string;
  _status?: ('draft' | 'published') | null;
}